refactor(cli): remove redundant shared re-export shims
