+++ src/cli/lib/manual/utils.js
@@ -8,7 +8,7 @@ import {
 } from "../shared-deps.js";
 import { formatDuration } from "../time-utils.js";
 import { formatBytes } from "../byte-format.js";
-import { isNonEmptyArray } from "../shared/array-utils.js";
+import { isNonEmptyArray } from "../../../shared/utils.js";
 import { writeManualFile } from "../manual-file-helpers.js";
 
 const MANUAL_REPO_ENV_VAR = "GML_MANUAL_REPO";
