+++ src/cli/lib/shared-deps.js
@@ -21,20 +21,20 @@ export {
     toNormalizedLowerCaseString,
     toPosixPath,
     toTrimmedString
-} from "./shared/utils.js";
+} from "../../shared/utils.js";
 
-export { normalizeStringList } from "./shared/utils/string.js";
+export { normalizeStringList } from "../../shared/utils/string.js";
 
 export {
     collectUniqueAncestorDirectories,
     resolveContainedRelativePath
-} from "./shared/path-utils.js";
+} from "../../shared/utils/path.js";
 
-export { escapeRegExp } from "./shared/regexp.js";
+export { escapeRegExp } from "../../shared/utils/regexp.js";
 
-export { getIdentifierText } from "./shared/ast.js";
+export { getIdentifierText } from "../../shared/ast.js";
 
 export {
     isAggregateErrorLike,
     isErrorLike
-} from "./shared/utils/capability-probes.js";
+} from "../../shared/utils/capability-probes.js";
