--- src/cli/lib/shared/array-utils.js
@@ -1,1 +1,0 @@
-export * from "../../../shared/array-utils.js";
