--- src/cli/lib/shared/fs-utils.js
@@ -1,1 +1,0 @@
-export * from "../../../shared/fs-utils.js";
