--- src/cli/lib/shared/number-utils.js
@@ -1,1 +1,0 @@
-export * from "../../../shared/number-utils.js";
