--- src/cli/lib/shared/path-utils.js
@@ -1,1 +1,0 @@
-export * from "../../../shared/path-utils.js";
