--- src/cli/lib/shared/regexp.js
@@ -1,1 +1,0 @@
-export * from "../../../shared/regexp.js";
