--- src/cli/lib/shared/string-utils.js
@@ -1,1 +1,0 @@
-export * from "../../../shared/string-utils.js";
