--- src/cli/lib/shared/utils.js
@@ -1,1 +1,0 @@
-export * from "../../../shared/utils.js";
