--- src/cli/lib/shared/utils/capability-probes.js
@@ -1,1 +1,0 @@
-export * from "../../../../shared/utils/capability-probes.js";
