--- src/cli/lib/shared/utils/string.js
@@ -1,1 +1,0 @@
-export * from "../../../../shared/utils/string.js";
